Beginning rebuild - see full commit for details
Objectives:
- Optimize assets to improve graphic ability
-- Removed biker and restored temporary 2D sprite
--- Will improve sprite in a later update, or restore a lower-def 3D
biker
-- Switched rendering mode to "Compatability" as mobile is not a target
platform
- Align song data to game time reactions
-- Added additional song configuration settings "lead-in" and
"travel-duration" to provide personalization to the user
-- Rebuilding object spawning system
--- Started with aligning objects to configured timings - want to follow
up with mid-song correction via timing polls
--- Beginning to redesign timing-based spawning - need to pull as much
processing out of main's _process(...) function as possible. Solution may be
to track a running index that indicates which object is next to spawn.
When the time at that index is reached, spawn object and increment
index.
--- Want to alter obstacle collision to provide time_scale lerping
instead of immediately shooting back to full speed
- Overhaul UI system
-- Create functions to modify labels without needing to track the label
title every time
-- Use organization elements to assure UI positions no matter the screen
size (to a limit)